services/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
asyncpg==0.30.0
psycopg[binary]==3.2.9
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
redis==5.2.1
aiokafka==0.12.0
httpx==0.28.1
requests==2.33.1
structlog==25.1.0
prometheus-fastapi-instrumentator==7.1.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-sqlalchemy==0.51b0
opentelemetry-instrumentation-redis==0.51b0
orjson==3.10.15
python-multipart==0.0.20
slowapi==0.1.9
pytest==8.3.5
pytest-asyncio==0.25.3
allure-pytest==2.13.5
ruff==0.9.9
